cleaned up the timer
